Alex

+++ src/app/services/product.service.ts
@@ -1,6 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { take } from 'rxjs';
+import { Categories } from '../data/categories';
 import { Price } from '../data/price';
 import { Product } from '../data/product';
 import { Sale } from '../data/sale';
@@ -31,9 +32,9 @@ export class ProductService {
     const newerDate = new Date();
     newerDate.setDate(new Date().getDate() - 80)
 
-    const defaultPrice = new Price(1,5,pastDate);
-    const newerPrice = new Price(2,4,newerDate);
-    const salePrice = new Sale(1,3,pastDate,futureDate);
+    const defaultPrice = new Price(1, 5, pastDate);
+    const newerPrice = new Price(2, 4, newerDate);
+    const salePrice = new Sale(1, 3, pastDate, futureDate);
 
     const availDate = new Date();
     availDate.setDate(new Date().getDate() - 100);
@@ -43,6 +44,7 @@ export class ProductService {
     product.scheduledPrices.push(defaultPrice);
     product.scheduledPrices.push(newerPrice);
     product.scheduledSales.push(salePrice);
+    product.scheduledMAPS.push(new Price(6,3.01,pastDate));
 
     this.products.push(product);
   }
@@ -64,10 +66,15 @@ export class ProductService {
       })
   }
 
-  createProduct(name: string, available: Date, description: string, price: number, imageURL: string) {
-    const defaultPrice = new Price(Math.random(), price, available);
+  createProduct(name: string, available: Date, description: string, price: number, imageURL: string, category: Categories | null, MAP: number) {
     const product = new Product(name, false, available, description, imageURL, [], [], [], [], []);
-    product.scheduledPrices.push(defaultPrice);
+    if (category)
+      product.categories.push(category);
+    product.scheduledPrices.push(new Price(Math.random(), price, available));
+    product.scheduledMAPS.push(new Price(Math.random(), MAP, available));
+
+    this.products.push(product);
+    this.creatingProduct = false;
     this.http.put("localhost:8080/products", product)
       .pipe(take(1))
       .subscribe({
@@ -81,17 +88,18 @@ export class ProductService {
       })
   }
 
-  deleteProduct(product: Product) {
-    this.http.delete("http://localhost:8080/products/${product.id}")
-      .pipe(take(1))
-      .subscribe({
-        next: () => {
-          this.updateProducts();
-        },
-        error: (error) => {
-          // print error to user
-        }
-      })
+  deleteProduct(product: Product | null) {
+    if (product)
+      this.http.delete("http://localhost:8080/products/${product.id}")
+        .pipe(take(1))
+        .subscribe({
+          next: () => {
+            this.updateProducts();
+          },
+          error: (error) => {
+            // print error to user
+          }
+        })
   }
 
   updateProduct(product: Product) {
@@ -113,7 +121,7 @@ export class ProductService {
 
     let currPriceDate = product.availability // start looping through dates at the date the item is available
     for (let price of product.scheduledPrices) {
-      if (price.date > currPriceDate && price.date <= todayDate) {
+      if (price.date >= currPriceDate && price.date <= todayDate) {
         defaultPrice = price.price; // update the product's default price for the current date
       }
     }
@@ -133,15 +141,28 @@ export class ProductService {
     return defaultPrice;
   }
 
+  getCurrentMAP(product: Product): number {
+    let defaultMAP = 0;
+    const todayDate = new Date();
+
+    let currMAPDate = product.availability // start looping through MAPS at the date the item is available
+    for (let MAP of product.scheduledMAPS) {
+      if (MAP.date >= currMAPDate && MAP.date <= todayDate) {
+        defaultMAP = MAP.price; // update the product's MAP for the current date
+      }
+    }
+    return defaultMAP;
+  }
+
   // returns the shipment the products are to be pulled from if purchased, returns null if product is out of stock
-  getCurrentShipment(product: Product) : Shipment | null{
-    let oldestShipment : Shipment | null = null;
-    let oldestDate : Date = new Date();
+  getCurrentShipment(product: Product): Shipment | null {
+    let oldestShipment: Shipment | null = null;
+    let oldestDate: Date = new Date();
     oldestDate.setDate(new Date().getDate() + Infinity);
 
     // find the oldest shipment with positive quantity
-    for(let shipment of product.shipments){
-      if(oldestDate > shipment.date && shipment.quantity > 0){
+    for (let shipment of product.shipments) {
+      if (oldestDate > shipment.date && shipment.quantity > 0) {
         oldestShipment = shipment;
       }
     }
@@ -149,9 +170,9 @@ export class ProductService {
   }
 
   // updates the product's shipments and returns true if successful, else returns false if the product is out of stock
-  attemptPurchase(product: Product){
+  attemptPurchase(product: Product) {
     let shipment = this.getCurrentShipment(product);
-    if(shipment){
+    if (shipment) {
       shipment.quantity--; // remove one of the products from the shipment
       this.updateProduct(product); // update the product now that its shipment has changed
       return true;
@@ -159,6 +180,34 @@ export class ProductService {
       // no shipment was found, so print error to user
       return false;
     }
+  }
+
+  findLossErrors(): Product[] {
+    let errorProducts: Product[] = []
+    for (let product of this.products) {
+      let price = this.getCurrentPrice(product)
+      let shipment = this.getCurrentShipment(product)
+      if (shipment) {
+        if (shipment.cost > price) {
+          errorProducts.push(product)
+        }
+      } else {
+        // the product is out of stock!
+      }
+    }
+    return errorProducts;
+  }
+
+  findMAPErrors(): Product[] {
+    let errorProducts: Product[] = []
+    for (let product of this.products) {
+      let price = this.getCurrentPrice(product)
+      let MAP = this.getCurrentMAP(product)
+      if (price < MAP) {
+        errorProducts.push(product)
+      }
+    }
+    return errorProducts;
   }
 
 }
